requirements.txt
Flask==0.10.1
Jinja2==2.7.2
MarkupSafe==0.18
Pillow==2.3.0
Werkzeug==0.9.4
argparse==1.2.1
gunicorn==18.0
itsdangerous==0.23
wsgiref==0.1.2
